[Fix] 본인인증 타입 관련 오류 수정 #131
